Trigger reconfiguration if verifier is created
Currently reconfiguration of the verifier is triggered only if a local
copy of the configuration is changed. In case the current verifier is
deleted and re-created, then the local copy of the configuration is not
changed, meaning that the code will not reconfigure the just-created
verifier.

The proposed change adds a trigger to the reconfiguration: if the
verifier has just been created, then perform the reconfiguration