Cccp 4555: Fixed cleanups for tempest runners (#40)
* Fixed missing openstack commands in template

* Improved cleanup for tempest runner

In our setup the domain deletion currently does not work, so removed it.
Fixed some other resource cleanups.
Added podman cleanups in the run scripts.

* Removed aggregate removal before the script

The aggregate removal often failed (no such aggregate), and caused the
script not to run. Removed the aggregate cleanup before the script. If
it occasionly fails after the script, it's fine and doesn't prevent
rally from running.